Port classify_image.py to master branch

+++ delta/imagery/tfrecord_conversions.py
@@ -7,6 +7,7 @@ import tensorflow as tf
 from tensorflow.python.framework.errors_impl import OutOfRangeError
 
 from delta.imagery import utilities
+from delta.imagery import image_reader
 from delta.imagery.sources import landsat
 from delta.imagery.sources import tfrecord
 from delta.imagery.sources import worldview
@@ -72,24 +73,37 @@ def _convert_image_to_tfrecord_landsat(input_path, work_folder):
     return (toa_paths, None)
 
 
-def _convert_image_to_tfrecord_worldview(input_path, work_folder):
+def _convert_image_to_tfrecord_worldview(input_path, work_folder, redo=False):
     """Convert one input WorldView file"""
 
     toa_path     = os.path.join(work_folder, 'toa.tif')
+    license_path = os.path.join(work_folder, 'license', 'NEXTVIEW.TXT')
     scene_info   = worldview.get_scene_info(input_path)
     bands_to_use = worldview.get_worldview_bands_to_use(scene_info['sensor'])
 
-    # Unzip the input file
-    print('Unzip file: ', input_path)
-    with zipfile.ZipFile(input_path, 'r') as zip_ref:
-        zip_ref.extractall(work_folder)
+    have_files = False
+    if utilities.file_is_good(license_path) and not redo:
+        (tif_path, meta_path) = worldview.get_files_from_unpack_folder(work_folder)
+        have_files = utilities.file_is_good(tif_path) and utilities.file_is_good(meta_path)
+
+    if have_files:
+        print('File ', input_path, ' is already unzipped.')
+    else:
+        # Unzip the input file
+        print('Unzip file: ', input_path)
+        with zipfile.ZipFile(input_path, 'r') as zip_ref:
+            zip_ref.extractall(work_folder)
+
 
     (tif_path, meta_path) = worldview.get_files_from_unpack_folder(work_folder)
 
     # TODO: Any benefit to passing in the tile size here?
-    print('TOA conversion...')
     # TODO get reflectance working!
-    worldview_toa.do_worldview_toa_conversion(tif_path, meta_path, toa_path, calc_reflectance=False)
+    if utilities.file_is_good(toa_path) and not redo:
+        print('TOA file ' + toa_path + ' already exists.')
+    else:
+        print('TOA conversion...')
+        worldview_toa.do_worldview_toa_conversion(tif_path, meta_path, toa_path, calc_reflectance=False)
     if not os.path.exists(toa_path):
         raise Exception('TOA conversion failed for: ', input_path)
     print('TOA conversion finished')
@@ -97,9 +111,15 @@ def _convert_image_to_tfrecord_worldview(input_path, work_folder):
     return ([toa_path], bands_to_use)
 
 
-def convert_image_to_tfrecord(input_path, output_paths, work_folder, tile_size, image_type):
-    """Convert a single image file (possibly compressed) of image_type into a single tfrecord
-       file at output_path.  work_folder is deleted if the conversion is successful."""
+def convert_image_to_tfrecord(input_path, output_paths, work_folder, tile_size, image_type,
+                              keep=False, redo=False, tile_overlap=0):
+    """Convert a single image file (possibly compressed) of image_type into TFRecord format.
+       If one output path is provided, all output data will be stored there compressed.  If
+       multiple output paths are provided, the output data will be divided randomly among those
+       files and they will be uncompressed.
+       work_folder is deleted if the conversion is successful."""
+
+    single_output = (len(output_paths) == 1) or (isinstance(output_paths, str))
 
     if not os.path.exists(work_folder):
         os.mkdir(work_folder)
@@ -114,9 +134,21 @@ def convert_image_to_tfrecord(input_path, output_paths, work_folder, tile_size, 
         raise Exception('Unrecognized image type: ' + image_type)
 
     # Generate the intermediate tiff files
-    tif_paths, bands_to_use = function(input_path, work_folder)
+    tif_paths, bands_to_use = function(input_path, work_folder) #, redo)
 
-    tfrecord.tiffs_to_tf_record(tif_paths, output_paths, tile_size, bands_to_use)
+    # Gather some image information which is hard to get later on
+    reader = image_reader.TiffReader()
+    reader.open_image(tif_paths[0]) # TODO: Check this indexing!
+    image_size = reader.image_size()
+    metadata   = reader.get_all_metadata()
 
-    # Remove all of the temporary files
-    os.system('rm -rf ' + work_folder)
+    if single_output and utilities.file_is_good(output_paths) and not redo:
+        print('Using existing TFRecord file: ' + str(output_paths))
+    else:
+        tfrecord.tiffs_to_tf_record(tif_paths, output_paths, tile_size, bands_to_use,
+                                    tile_overlap)
+
+    if not keep: # Remove all of the temporary files
+        os.system('rm -rf ' + work_folder)
+
+    return (image_size, metadata)
